+++ scripts/_settings.mjs
@@ -3,19 +3,37 @@ import { MODULE } from './_module.mjs';
 
 // FOUNDRY HOOKS -> SETUP
 Hooks.once('setup', () => {
-	// SET MODULE SETTINGS
-	// SET MODULE SETTINGS
-	MODULE.setting('register', 'trackedChangelogs', {
-		type: Object,
-		default: {},
+	// SET MODULE MIGRATE SETTINGS
+	MODULE.setting('register', 'clientMigratedVersion', {
+		type: String,
+		default: "0.0.0",
 		scope: 'world',
 		config: false
 	});
-	
+	MODULE.setting('register', 'worldMigratedVersion', {
+		type: String,
+		default: "0.0.0",
+		scope: 'world',
+		config: false
+	});
+
+	// SET MODULE SETTINGS
 	MODULE.setting('register', 'showNewChangelogsOnLoad', {
 		type: Boolean,
 		default: true,
 		scope: 'world',
+	});
+	MODULE.setting('register', 'trackedChangelogs', {
+		type: Object,
+		default: {},
+		scope: 'client',
+		config: false
+	});
+	MODULE.setting('register', 'presets', {
+		type: Object,
+		default: {},
+		scope: 'client',
+		config: false
 	});
 });