+++ scripts/init.mjs
@@ -4,13 +4,9 @@ import { MODULE } from './_module.mjs';
 // IMPORT SETTINGS -> Settings Register on Hooks.Setup
 import './_settings.mjs';
 
-//import * as packages from '/common/packages.mjs';
-import { MMP } from './module.mjs';
 
-Hooks.once('socketlib.ready', () => {
-	MODULE.debug('SOCKETLIB Ready - SOCKET'); // WONT REGISTER CAUSE CALL HAPPENS WAY TO EARLY
-	MMP.registerSocketLib();
-});
+import { MIGRATE } from './_migrate.mjs';
+import { MMP } from './module.mjs';
 
 /* ─────────────── ⋆⋅☆⋅⋆ ─────────────── */
 // 🧙 DEVELOPER MODE HOOKS -> devModeReady
@@ -18,23 +14,36 @@ Hooks.once('socketlib.ready', () => {
 Hooks.once('devModeReady', ({ registerPackageDebugFlag }) => {
     registerPackageDebugFlag(MODULE.ID, 'level', {
 		choiceLabelOverrides: {
-			NONE: 0,
-			ERROR: 1,
-			WARN: 2,
-			DEBUG: 3,
-			INFO: 4,
-			ALL: 5,
+			0: 'NONE',
+			1: 'ERROR',
+			2: 'WARN',
+			3: 'DEBUG',
+			4: 'INFO',
+			5: 'ALL'
 		}
 	});
+});
+
+/* ─────────────── ⋆⋅☆⋅⋆ ─────────────── */
+// socketlib HOOKS -> socketlib.ready
+/* ─────────────── ⋆⋅☆⋅⋆ ─────────────── */
+Hooks.once('socketlib.ready', () => {
+	MODULE.debug('SOCKETLIB Ready - SOCKET'); // WONT REGISTER CAUSE CALL HAPPENS WAY TO EARLY
+	MMP.registerSocketLib();
 });
 
 /* ─────────────── ⋆⋅☆⋅⋆ ─────────────── */
 // FOUNDRY HOOKS -> READY
 /* ─────────────── ⋆⋅☆⋅⋆ ─────────────── */
 Hooks.once('ready', async () => {
+	await MIGRATE.init();
+	
 	MMP.init();
 });
 
-Hooks.on('renderSidebarTab', MMP.updateSettingsTab)
+/* ─────────────── ⋆⋅☆⋅⋆ ─────────────── */
+// FOUNDRY HOOKS -> MODULE FUNCTIONS
+/* ─────────────── ⋆⋅☆⋅⋆ ─────────────── */
+Hooks.on('renderSidebarTab', MMP.updateSettingsTab);
 Hooks.on('renderModuleManagement', MMP.renderModuleManagement);
 Hooks.on('renderSettingsConfig', MMP.renderSettingsConfig);